requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=2.0.0
pinecone-client==3.0.0
PyPDF2==3.0.1
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0